requirements.txt
Jinja2~=3.1.4
PyYAML~=6.0.2
# Email Exporter dependencies
python-dotenv~=1.0.0
beautifulsoup4~=4.12.0
html2text~=2020.1.16